* 'master' of github.com:linkedin/gobblin:
  Fix division by long in PolicyChecker
  Set task state to failed when commit to staging directory fails
  Addressed comments for PR #248
  Refactor Writer and Publisher, and add a Util class for publishing Kafka audit counts
  Set (optional) group in commit only if the staging file exists
  Moved code dealing with general file level arrributes to FsDataWriter
  Gobblin-data-management uses Properties instead of azkaba Props (for easier compatibility).

+++ gobblin-core/src/main/java/gobblin/policies/count/RowCountRangePolicy.java
@@ -31,7 +31,7 @@ public class RowCountRangePolicy extends TaskLevelPolicy {
 
   @Override
   public Result executePolicy() {
-    if (Math.abs((this.rowsWritten - this.rowsRead) / this.rowsRead) <= this.range) {
+    if (Math.abs((this.rowsWritten - this.rowsRead) / (double)this.rowsRead) <= this.range) {
       return Result.PASSED;
     } else {
       return Result.FAILED;
